requirements.txt
flask
streamlit
google-genai
requests
